Ajustar o layout de admin

+++ app/helpers/application_helper.rb
@@ -1,9 +1,6 @@
 # frozen_string_literal: true
 
 module ApplicationHelper
-  def logout_admin
-    logout_builder('Fazer logout [Administrador]', destroy_admin_session_path)
-  end
 
   def logout_user
     logout_builder('Fazer logout [Usuário]', destroy_user_session_path)
@@ -11,7 +8,18 @@ module ApplicationHelper
 
   private
 
+  def logout_with_icon(text,route)
+    link_to route, method: :delete do
+      yield
+      text
+    end
+  end
+
   def logout_builder(link_text, route)
-    link_to link_text, route, method: :delete
+    if block_given?
+      logout_with_icon(link_text,route) { block.call }
+    else
+      link_to link_text, route, method: :delete
+    end
   end
 end
